[vision/monitor] revamp monitoring intro section
Assume that monitoring state is stored in the resource and
job databases, leveraging their extensibility and persistence features.
Since all the state is in there, assume the interface for accessing
monitoring data is the resource management API, and the tools for
doing that (a web interface for example) belong under resource
management.

The "log database" and the "enterprise monitoring system" (ops/sys admin
monitoring displays) are outsourced, and the NGRM activity will be to
develop interfaces rather than implementation.

Included the CIFTS FTB-API in monitoring.