Use pdflatex instead of latex/dvips/pstopdf
This improves some of the graphics, as the pdf output is able
to handle transparency better. The conversion of eps to pdf is
handled in the background by the epstopdf package. This slows
down the build, but could be remedied by doing a onetime
conversion of eps->pdf.

I also added a \graphicspath so that we don't have to reference
../fig/ in every \includegraphics, but I only updated some of
the references.