[zmq-broker] Tree network routed request-response
Arrange for router socket to listen on tree input uri as well
as the inproc plugin address.  Add a dealer socket for tree
output that the broker can route requests to.  Now it is possible
for requests to be routed upstream using 0MQ extended req-rep
addressing, and for replies to be sent back downstream.

"Smart gateway" routing:  If a message is sent to a plugin (tag)
that doesn't match locally, it is sent upstream.  Services can
transparently be taken off of leaf nodes (e.g. for noise or memory
constraints), or implemented trivially at the root of the tree,
without the client code knowing the difference.

"distributed plugin" routing: A plugin can use the reduction tree
as before by sending request messages to the router socket.  Requests
with a hopcount of 1 (local) that match the sender's id (tag/plugin name)
are forwarded upstream to the parent.  Replies are now a possibility.