[zmq-broker] get events working in api plugin
Manage per-client event subscriptions and subscribe to the
union of all the client subscriptions via the broker.
Change the API interface to allow events to be sent, received,
subscribed and unsubscribed to.  Update cmbutil to exercise
these API calls.

Add 'stats' function, similar to ping, that is automatically handled
by the default poll loop, if the plugin does not provide one.

Rework barriersrv, livesrv to work with new broker sockets.
Untested so far.