zmq-broker: abstract "treeroot" test somewhat
Plugins would test conf->rank == 0 to determine if they are
at the root of the reduction tree or not.  Eventually the root
might move to a different rank so abstract this test behind
plugin_treeroot () accessor.

Also, cmbd makes routing decisions based on existence of upstream
and downstream plumbing.  Ensure that proper plumbing exists for
rank 0 and non-rank 0.