zmq-broker: improve barrier implementation
Triggering clients directly with barrier completion event was racy.
We subscribed to the barrier exit event, entered the barrier, and then
blocked waiting for the event.  This is racy in zmq 3 where subscription
requests are messages with latency instead of just subscriber end
filters.  An event could be handled before the subscription request.

Instead, the barrier plugin tracks clients and sends them regular
response messages when the barrier terminates (still signalled by
event).  In addition, clients that disconnect premature are tracked,
and this now can cause a barrier abort.