zmq-broker: transmit debug logs on fault event
At the root of the tree, send logs to log.[ch], which can
be configured for file, syslog, or stderr.  Set pepe.conf files
to log to cmbd.log.

Only logs with a priority less than a threshold are
persisted/forwarded to root, but all priorites go to the local
circular buffer.

If event.fault.<facility> is received, send debug logs from
the circular buffer that match <facility> up to the root and
persist them.