zmq-broker: add confsrv alternate k-v store plugin
Experimental 'conf' plugin (ostensibly for configuration params).

Master store is at the root of the tree with integer version number.
All changes are to the master copy;  a "put" request recurses up the
tree if not initiated at the root.  Gets can be satisfied out of local
cache if version is up to date, otherwise they recurse until satisfied.
When version is incremented (on every put) an event is generated which
invalidates the all caches of older versions of store store.