zmq-broker: fix bug in conf caching
We initialize the conf cache during bootstrap which generates
and update event, but this event may not be received by all clients
since we are just bootstrapping.  If a response from upstream includes
a version number newer than the current cache, drop the cache and
update the version before storing the new value in the cache.