zmq-broker: avoid buggy czmq zmsg_addstr
We use this in cmb_msg_encode.  I get a segfault encoding large json
strings (larger than 256 bytes) down in zmsg_addstr.
Looking at the source, it appears that the version of czmq we have
packaged from EPEL (1.3.2) has incorrect stdargs handing in the case
where the buffer is >256 bytes and has to be reallocated.

Avoid it and use zmsg_addmem which is more efficient in this case
anyway.