zmq-broker/livesrv: don't track prev parent state
Upon receipt of a live.hello message, we begin monitoring
nodes that have chosen us as their parent.  This occurs when a
node's primary parent is declared down, and it reacts by
connecting to its secondary, if any.

We used to monitor the state of the primary parent and stop montoring
children whose primary came back, presuming that the node would
reparent.  Disable this behavior, in anticipation of the topology
being updated in response to reparenting.