zmq-broker: add 4ary and 48ary pepe configs
If we launch cmb as a k-ary tree, the minimum height for 100k nodes
would be h=3 at k=48.  For experiments at lower scales it may be useful
to play with those values of h and k.  E.g. at 128 nodes, try k=48 (h=2)
to experience what might be full scale fan-in (reduction factor) and
k=4 (h=4) to experience what might be full scale RTT.