zmq-broker: restructure logging to avoid cmb.h includes
Changed logging to use "level" instead of "priority", which is
more in line with syslog terminology), and to use the syslog.h
values for that, as an integer, not our own enum.

This allows various .c files to avoid including cmb.h just to
get the level values.  Instead they should include log.h which
in turn includes <syslog.h>.