apisrv: on fd write error, defer destruction
There was a bug repeated in three places where a write error on
API client fd would, in an attempt to end the client's poller
in the zloop reactor, remove another client instead.  Since this
code was trying to end a poller in the handler of another poller,
the code was awkward anyway.  Just remove this code and expect that
there will be a fd read error event in the client's poller callback,
where it can be removed once, and perhaps more safely since this
is likely to be the expected czmq zloop usage.