pmi: add PMI_USE_KVS_CACHE attempt at performance
If PMI_USE_KVS_CACHE=1 in the job's environment, libpmi will cache
the job's kvs namespace on first KVS_Get after a KVS_Barrier.
In theory this should eliminate many RTT's between application and CMB
when mvapich2 attempts to share some keys globally; however in practice
performance is worse and we see some failure modes such as PMI_KVS_Get
failing with errno set to EPROTO.

I'm not sure if the failures are caused by code introduced in this
commit, or if it is exercising the underlying KVS code in a way that
exposes a problem.  This and the unexpected performance results
require investigation.