kvssrv: add watch capability
The interface is call cmb_kvs_get with KVS_FLAGS_WATCH and receive
a reply like a normal get.  Then call cmb_kvs_get with KVS_FLAGS_RCVONLY
and the call will block until the requested key changes.

(Actually this will get called on every commit whether the key changes
or not, so it is not quite there yet).