util: remove lz4 support
This seems not to help performance, and as we intend to ecrypt messages
in the future, we should be concerned about creating side channel
vulnerabilities by sending encrypt(compress(message)) over the wire
(pointed out by Chris Dunlap, LLNL).

More info on the security impact:
"Compression and Information Leakage of Plaintext", by Kelsey, 2002.
http://www.iacr.org/cryptodb/archive/2002/FSE/3091/3091.pdf