plugin: handle unexpected reply messages
When processing a request-reply synchronously within a plugin,
e.g. calling the kvs functions in plugin.c, if a reply to another
request arrives, e.g. one being handled asynchronously in the plugin,
or an asynchronous watch reply handled internally, put that reply
on a list and process it when the plugin callback that made the
synchronous request returns.

This fixes a bug in the KVS code where a watch reply is generated during
a commit, but before the commit reply is sent.  A plugin executing
plugin_kvs_commit () would receive the watch reply and discard it,
and the watch callback in the plugin would not be called.