kvs: tighten up directory handling
Return error (ENOENT currently) if KVS_GET_DIRECTORY is requested
but value is not a directory, or if value is a directory and
KVS_GET_DIRECTORY is not requested.

Return EINVAL if KVS_GET_DIRECTORY or KVS_GET_DEEP is passed to
one of the simple type kvs_get_<type> accessors.

Add "get_dir" and "get"/"put" (raw JSON) operators to tkvs.c.