kvs: add flags to kvs_get_dir that affect caching
Set flags = 0 if kvsdir_get* and kvsdir_is* functions should
always make kvs.get requests to obtain the latest values for keys
relative to the kvsdir object.

Set flags = KVS_GET_FILEVAL if non-directory values should be cached
in the kvsdir object when it is initially obtained.

Set flags = KVS_GET_DIRVAL if directory values should be cached
in the kvsdir object when it is initially obtained.  This is recursive.

Setting both flags caches the entire namespace relative to the kvsdir
object in the object.