kvs: fix kvs error handling code
Change the way plugin_request handles a response containing only the
integer 'errnum'.  We used to return a NULL response with errno set if
errnum was nonzero.  However, in several places (e.g. kvs.put) an errnum
= 0 response indicates success.  Modify plugin_request so that a NULL
return with errno = 0 is not an error.

In kvscli, remove 'errnum' handling code from all the places this
function is used.