kvs: handle symlinks in kvsdir_get functions
Implement kvsdir_get_symlink.

If symlink encountered in path traversal of kvsdir object, fall back
to a server request and let it resolve the symlink.

FIXME: in this simple implementation, the value of a symlink that
may be stored in the kvsdir_t is never used.  Instead, if a symlink
is encountered either in path traversal or as the final path component,
the code falls back to a kvs_get* code path, ignoring the cached value.
This should probably be fixed in a future implementation.