kvs: drop flags from kvs_get_dir(), kvs_watch_dir()
Eliminate flags for requesting "cache" in the kvsdir_t objects.
All kvsdir_get() functions now just make a call to the server
with a relative path rather than trying to pull values out of cache.

The caching was complicated, it was likely a premature optimization,
and it had some inconsistencies that were counterintuitive, such
as in symlink handling (symlink value never read from cache).