flux_t: eliminiate handle ops for rank,size,treeroot
Have cmbd supply this information (which in the future may be
dynamic) rather than storing static values in the flux handle.

Note this requires the previous commit to avoid deadlock in
KVS initialiation (kvssrv_init() calls flux_treeroot() but we
were starting and sending puts/commits to it before beginning the
cmbd poll loop, which is required to answer the treeroot query.