ping: add flux_ping(), restore cmbutil -p and -x
Ping message parsing was hardwired in cmbutil.c.
Move this to new flux_ping() function.

Along the way, change cmb_msg_match() so that a N! style
address is stripped before comparison if it appears either in
the message or the tag.  Previously it could only be in the message.
This allows 'cmbutil -p N!whatever' to work again.

Restore the cmbutil -x (stats) option.  Dod not add a new flux_*
function for stats as for now it suffices to just convert the returned
JSON to a string and print it.