flux-reactor: send FLUX_MSGTYPE across API socket
To implement the flux-reactor for the API socket, we need to be able
to handle different message types that arrive on the shared file
descriptor.  Send the message type along and implement message queues
in the handle implementation so that messages of one type arriving when
another is expected don't have to be dropped.