dlua: Compile 'kvs' module into flux lua module
Moving the kvsdir methods into the kvsdir object obviates the
need for compiling kvs.so as a separate module. This will make things
easier when linking flux Lua code to other applications, so do this
for now.

Note: Placing kvsdir methods in the kvsdir object means that
entries in the kvs with those names can't be accessed. These include
watch, watchdir, keys, etc.