wreck: Use private env for LWJ tasks
Instead of using the environment of wrexecd to contain the environ
of LWJ tasks, use a seperate environ array for the job tasks.

This necessitated a new Lua interface to the environment, exported
via wreck.environ to Lua scripts. This environment can be indexed
and set like a normal Lua table, e.g.:

 local env = wreck.environ
 env.PATH = env.PATH .. ":/opt/newpath/bin"