flux-util: Check for subcommands in same dir as flux util
If a subcommand is not found in libexecdir/$cmd, also check
in flux_exe_dir/flux-$cmd. This allows scripts to be dropped
directly into $builddir/flux-NAME for testing, instead of
requiring a rebuild or 'make install' to copy the source
of a script to libexec/.

For example, now 'flux wreckrun' will directly invoke the
flux-wreckrun lua script committed to top srcdir.

(Note this breaks if we ever support a builddir separate from
srcdir, but we can do a proper fix for all this when we rework
the build system)