wreck: flux-wreckrun: Change to support scheduler work
This commit changes the sense of -n option to flux-wreckrun to
mean number of total tasks, and sets 'ntasks' in the jobid reqeust
field instead of nprocs.

Additionally, an -i, --now flag is added to force invocation of
the "run" event from wreckrun itself. By default this is disabled
and wreckrun will wait for the scheduler to issue the run event.

Finally, a -N, --tasks-per-node flag is added. For the old behavior
of wreckrun, use both -N and -i, e.g.

 flux wreckrun -N 2 -i hostname

will immediately run 2 copies of hostname across all hosts.