zio: add kz interface to zutil for KVS streams
kz provides an abstraction for storing streams in the KVS.
The zio json object encapsulatuion is used for KVS values.
A stream is represented by a directory containing numerically
sequenced keys, one for each json object in the stream.

kz_open()
kz_close()
kz_flush()
kz_put()
kz_get()
kz_set_ready_cb()

Both blocking and non-blocking read interfaces are provided.
If a kz_t stream is opened with the KZ_FLAGS_NONBLOCK clear, then
a kz_get() on the stream will block until data is available
or an object with the EOF flag set is written.  If on the other
hand that flag is set, kz_get() will return EAGAIN if no data is
available.  kz_set_ready_cb() can be used to register a function
that will be called when data is avilable to be read with kz_get().
A kz_t stream can thus be integrated into a flux reactor loop
with other event sources.

In the current code, a kvs_commit() is issued after every kz_put()
and once per kz_flush() and kz_close() if the kz_t stream is open
for writing.  In the future the frequency of commits may be controlled
by flags.

In conjunction with the addition of this module, functions were exposed
in zio for encoding/decoding zio json:

zio_json_decode()
zio_json_encode()