kvs: restructure commit internally, fence broken
Got rid of writeback queue and flush message.
Instead, a kvs.put creates a commit hash for the sender that captures
metadata (key = dirent).  The data (e.g. large values put in the store)
is written separately in a writeback style as before.

When kvs.commit is called, the commit stalls while any 'store'
operations are in progress, then pushes the commit up one level towards
the root, repeating until the root is reached.

A big change is that a client disconnecting before issuing kvs_commit
will lose those commits.  I had to rework all the tests for this.
N.B. We could add a "commit on disconnect" option if necessary.

Also, a commit with nothing pending is now really a no-op, and one that
can be handled without involving the master.

What works:
- cd test; make check
- cd test; make check3

What doesn't
- kvs_fence