test/tkvswatch: relax test requiring all updates be seen
A kvs_watch callback is not guaranteed to be called for every update
since intermediate commits (root values) can be skipped by a client
if it receives a newer root reference out of band, for example
in a commit response.

Therefore relax this test to ensure that only the last update
(the "latest") is received.