kvs: replace 'version' key with kvs.sync,getroot
Get rid of the special 'version' key because it made it impossible
for a commit to be a no-op (the root directory SHA1 would always
change because the version was incremented).

Do not increment the store version if it hasn't changed after a commit.

Reimplement get_version and wait_version as kvs.getroot and kvs.sync.