kvs: include root directory in event.kvs.setroot
Since the root directory is generally faulted in immediately
on all slave nodes right after every update (due to watched
keys), we can eliminate this transaction by sending a copy
of the directory in the event message.

N.B. this may not be a good idea if the root directory grows
large.