mon: add monitoring plugin
monsrv - depending on the keys found in conf.mon.source.*, on every
session heartbeat, generate a JSON object that is placed in
mon.<rank>.<epoch>.

flux-mon - list/add/del monitoring sources.

At the moment the only "source" type is "rpc" which arranges for
the monsrv thread to make a request to another plugin, and incorporate
the JSON response into its sample for each epoch.  For example

  ./flux mon add cmb-rusage rpc cmb.rusage

adds a source which incorporates the JSON response from the cmb.rsuage
request into the epoch's JSON object under the key "cmb-rusage".
On node 42, heartbeat epoch 100, the aggregate JSON object would be
stored in the key "mon.42.100".