auth: use CURVE instead of PLAIN
Create two public-private key pairs, one for "client" and one for
"server".  Use curve authentication (client to server) when wiring up
to parent.  This is just for testing - presumably in a proper setup,
each rank would have its own key pair, and there would be a mechanism
to get the public key of a server before issuing a connect.

However I think we can get at least a little protection by sharing
tese two keys in NFS among all the ranks of a session.  In theory
data is encrypted and connections from outside the job would be denied.

I only added the auth code to the dealer/router sockets.  I did not have
time to experiment with PUB/SUB, especially over PGM.

Note: sockets with auth configured that are simultaneously bound to
TCP and "inproc" seem to require auth from TCP but not inproc.
I did not check whether auth is similarly bypassed on the "ipc" transport.