cmbd: do not log runt epgm messages at startup
A few zero size messages are received on our epgm socket.
Also sometimes we see message parts without the rest of the message.
This seems to happen at only startup.

This has been complained about on zeromq-dev but does not appear
to have been resolved?  Verified that we see them on 'master'
and stop logging them in the new cmbd code (just silently discard
as before).