event: more flexible "local" socket endpoints
Instead of hardwiring an ipc socket for an event nexus per node,
arrange for the event module to open an inproc socket which the
cmbd interrogates with event.geturi.

Allow the event module to not be loaded, and then for the request
to recurse up the cmbd request tree until it is found.  Pass the
client pid and hostname in on the event.geturi request and if a
different process or different node is requesting, bind to a
random ipc:// or tcp:// endpoint as appropriate and return that.

This handles three important use cases well:
1) one cmbd per node using epgm://  (all our pepe configs)
   Use low-overhead inproc:// between cmbd and eventsrv, (not ipc://)
2) multiple cmbds on one node  (screen-3 config)
   Load event module on on treeroot only, cmbds subscribe to ipc://
3) lightweight compute node
   Load event module from treeroot to leaf level - 1.
   Leaf cmbds subscribe to parent's tcp://