zmsg_route_str: output
Change the zmsg_route_str() format so that routing frames are
represented as "sender!hop!hop!hop[!...]destination" with uuids
shortened to the first 5 characters.  This order is the reverse
of how it appears in the zmsg and how it was displayed before,
but it is more intuitive since it corresponds to the way pings
are constructed: node!service.

This affects the string displayed in ping responses.

Also changed zmsg_dump_abbrev() to use this form instead of the
previous one with routes separated by vertical bars.