cmbd: handle SIGCHLD gracefully instead of terminating zloop
When running external processes cmbd may recieve SIGCHLD as a result
of these processes exiting or daemonizing. The current code terminates
the reactor loop on any signal, but this is inappropriate for an
exiting child.

Instead, in case of SIGCHLD just clean up any child processes with
waitpid(2), but do not terminate the reactor loop.