modctl: add lsmod reduction and seq number
When we update conf.modctl.modules, we also update conf.modctl.seq.
Then we watch this in modctl instead of the conf.modctl directory.
After syncing loaded modules with the list of modules in the kvs,
we do a cmb.lsmod and push the results through a reduction handle
towards the root.

The idea is then a KVS value will contain the aggregated lsmod results
for the whole session.  Each module would have a nodelist showing which
nodes have it loaded.  Only the results from a particular sequence is
aggregated.