librdl: fix stack bug in lua_rdl_resource_getfield
lua_getfield() leaves the table from which getfield is called on the
stack. Sometimes this doesn't matter, but when building up arguments
to a function or other uses, it will unexpectedly add extra items on
the stack.

For example, with lua_rdl_resource_getfield(), the expectation is
that after the call the stack has only one extra entry -- the value
of the field requested -- at the top of the stack. This was not the
case in the current code.

This patch uses lua_replace(L, -2) to replace the resource table in
the stack with the returned field.