sched: newlwj_cb: avoid race during lwj creation
newlwj_cb() is called when a new lwj is detected.
It installs a lwjstate_cb() watching the lwj's .state,
then adds it to the p_queue.  lwjstate_cb() is always
called once in the context of the kvs_watch() call.
If the state was not yet available (lwj in process of
creation) no problem.  If the state was available,
then lwjstate_cb() would try to find_lwj() to create
an event, and would fail to find it.

Fix: add lwj to the p_queue before calling
rej_lwj_state_hdlr().