live: track peer idle time for plugins
Keep a hash by uuid/sender id for tracking peer idle times.
Record the heartbeat epoch when a peer last communicated with the cmbd.

request_cb - if hopcount > 0 (meaning sender is not the cmbd itself),
update the peer entry for 'lasthop'.  This will either be a plugin uuid,
or a neighbor node in the request overlay.

plugins_cb - update the peer entry for the plugin uuid.