live: add flux_failover(), flux_recover()
Add dumb instructions to the cmbd to:

failover - use a new parent for new requests, while listening for
responses on both new and old parents (and old continually trying
to reconnect as zeromq sockets do if it has actually lost the connection).

recover - go back to using the original (configured) parent for new
requests, while listening for responses on both new and old parents.
For now we leave the old parent connected.

The flux-peer utility currently can drive these directly.