live: split comms modules/peer sockets
With new resiliency code, it became apparent that it was easy
to starve out downstream peers on the request socket by creating
a large inproc:// load using flux-kvstorture.

This would trigger inappropriate transitions of downstream (idle
but healthy) peers to 'slow' or 'fail' state.

Split the plugins and the downstream peers onto two separate
request sockets.  This seems to have addressed the starvation issue
at least for limited testing.