rdl: Allow table to be passed to tag()
Allow a table of tags to be passed to MemStore:tag() function.
If the `tag' argument is found to be a table, assume it is a table
of key/value tags, possibly including a list of boolean tags in
the array part.

Adds a function to convert a table of tags to all key-value pairs.
That is, tags in the array part are converted to 'tag' = true.