rdl: Properly handle attaching a new hierarchy at non-root URIs
In MemStore:hierarchy_put() non-root URIs were not working properly,
that is, new hierarchies could only be attached at the root.

This patch fixes hierarchy_put() such that attaching a new sub-hierarchy
to an existing tree will work, e.g.

 hierarchy_put ("default:/foo", node)

attaches the hierarchy 'node' as a child of resource "foo" in the
default hierarchy.