pmi-boot: handle pmi cliques with event relay
When running multiple cmbds per node and bootstrapping with PMI,
select a node in each "clique" (group of processes on one node)
to subscribe to the epgm:// event socket and relay it to others
via an ipc:// socket.

This works around the fact that one one process per node can
subscribe to a given epgm:// endpoint.

Caveat: liveness module tells children that their parent has gone
down (so they can reconnect) using events.  If the parent also happens
to be the event relay, those children will not be notified.