emulator is now relatively stable
Fixed several bugs, including a pernicious race condition with job
creation in the kvs also setup useful printing to LOG_INFO to enable
post-processing/analysis

Note:  This commit was split between ngrm and flux-core repos.
The job.c workaround is #ifdefed out as it makes an rpc to the simulator
scheduler.  -garlick 2014-10-20