pmi-boot: add shutdown event
When the rank0 shell terminates, or on receipt of fatal signals,
send out a shutdown event to the session.  On receipt of this event,
install a timer for the (received) grace period, then exit when
the timer expires.

This allows the session to exist long enough to propagate the
shutdown signal, but it is really just a placeholder for proper
synchronous shutdown, which is tricky and seems to require modules
to be unloaded in reverse dependency order, and also leaves-to-root
hierarchical order.  Additional considerations: allowing modules
time to commit to stable storage, shrinking to a minimum size,
then allowing state to be "reaped", etc.