autoprep: re-enable zio utility
Drop lsd_fatal.c from libutil and compile liblsd so that it uses
its default error handling and does not depend on lsd_nomem_error()
etc.  The way cbuf is being used by zio, and list is being used by optparse,
it would appear that NULL returns and so forth are being handled in
those places.

Wreck on the other hand was building wrexecd with its own lsd_nomem_error
handler, but it returned NULL so I think we are OK.

Also: reorder and drop some superfluous library linking