log: simplify libflux/flog.c
Add flux_log_set_redirect() which causes new log messages to be sent
to libutil/log.c instead of being sent as a request on the handle.
This short circuits an awkward situation in cmbd.c where calling
flux_log() on rank 0 would generate a message that had to be looped
back to a handler in the same thread.

Move all JSON encoding/decoding of Flux log messages to libflux/flog.c
and add flux_log_zmsg() which will respect the redirect flag above.
Thus the handler for cmb.log in cmbd.c just has to set the redirect
flag and call when a request is received.

Change the JSON encoding to
- drop the integer 'count' used in reduction as we are not doing
  that right now
- change the string "source" to integer "rank"