wreck: Update name of wrexec module
cmbd fails to load the wrexec.so plugin on the command line,
presumably because the registered plugin name "rexec" does not match
the name of the DSO (wrexec.so).

Probably the wreck service should be called "wrexec" anyway, to
differentiate it from a future rexec service, so this patch updates
all references to "rexec" to "wrexec".