lua: move possibly conflicting named lua modules to flux-lua
For the helper Lua modules 'timer' and 'alt_getopt' it may
be dangerous to install in the default Lua path. Especially,
alt_getopt may conflict with the system alt_getopt since
this exists as an external package.

For the time being, install these modules into subdirectory
'flux-lua' to indicate they are helper lua libraries for flux,
and require them explicitly by prepending 'flux-core.' to the
require line.

Note we also move these libs to flux-core/ subdirectory in
the source tree to not only indicate the destination of these
modules, but so that require will "just work" from builddir.