doc: don't define dist_man_MANS with substititon
automake is getting confused when dist_man_MANS is defined
as a pattern substitution on a list of .adoc files.
(make install installed nothing)  Reverse the def so that
$(ADOC_FILES) is defined based on pattern substitution.
Then we have to base the 'make check' rule on $(MAN1_LIST)
and deal with 'roff files in the spellcheck script.

This is better anyway, since aspell has a 'nroff' mode (-n),
and we were just winging it checking the adoc files without a
filter mode.