flux-start: misc tidy up
If --nnodes or --partition is used, launch with slurm.

If --nnodes is specified (implying slurm), and the number of nodes
exceeds the requested size (or --size is left unset), set the size
to the number of nodes.

Replace the --method option with a --slurm option.

Use the full path to srun since we are using execv() not execvp().

Add srun --pty option, and rename srun --nnodes to --nodes