flux-start: use POSIXLY_CORRECT when processing args with getopt
We don't want to parse flux-start `command` args as our own, so
be sure to set the POSIXLY_CORRECT flag for getopt_long() so
that it stops processing at the first non-option arg. That way
commands with arguments are processed unharmed. (This is done
by prepending '+' to shortopts in getopt_long(3).

For example, before this commit, flux-start eradicates all options
to wreckrun in this test:

 flux start -N4 flux wreckrun -N4 hostname

after the commit, wreckrun args should be preserved.