test: get spellcheck working with newer automake
Putting a command in TEST_ENVIRONMENT and expecting it
to iterate over a list of arguments in TEST is fragile.
In newer automakes, other things are appended to TEST_ENVIRONMENT
and this breaks.

Modify the man page spell checker so that it is one test
called 'spellcheck' that checks all the *.[1-9] files it can glob.