cmbd: module name containing "/" is loaded by path
If a module name is specified to cmbd with a / in it, presume that
this is the path to a shared object file in module format and
treat failure to load it as a module as a fatal error.

The old scheme would try to load 'barrier' as a module if it existed
in the current working directory.  Now we would require ./barrier
if that is desired.