config: rework flux_conf_* API
Create a flux_conf_t class which hides the czmq::zconfig API behind
a much simpler one.  Users of the class will see a flat list of
keys, albeit keys with path separators that allow them to be mapped
into the hierarchical key space of the KVS or zconfig interchangeably.

Since not all users of the flux_conf_t class will want to link
with libkvs.la and its dependencies, the class provides suitable
interfaces for external code to load/store the cached config in the KVS.
(This will be later added to libkvs.la)

The API provides
 - create/destroy
 - set_directory/get_directory to get/set the config directory
   (a default of $HOME/.flux is used if set is not called)
 - load/save/clear to manipulate an internal cache relative to config file
 - get/put values in the internal cache
 - iterator to generate and walk an ordered list of keys

Unit tests are a part of this commit.