libflux/security.c: fix return code test on zmsg_encode()
zmsg_encode() returns size_t, so it is invalid to test for a return
value less than zero here. In fact, zmsg_encode() doesn't appear
to return an error at all -- even if malloc(3) fails it will return
the buffer_size that was computed from the zmsg.

For now, silence warnings from clang by detecting a return code of 0
from zmsg_encode() as an error. Also, errno isn't set, so don't
try to use it in the error message.